Cleanup: Move command modules into subpackge
